refactor orders page to server-side fetch
